Folders renaming and structure now updated (#191)
