+++ client/src/app/components/courses/course-actions/course-edit/course-edit.component.scss
