--- client/src/app/components/courses/pages/list/courses-list.component.scss
@@ -1,131 +1,0 @@
-.courses {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 1rem;
-    justify-content: center;
-    margin-top: 2rem;
-}
-
-.courses-container {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-
-    .div-title-course {
-        width: 100%;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        margin-top: 8.5rem;
-        margin-bottom: 1.5rem;
-
-        span {
-            font-size: 35px;
-            color: black;
-            font-weight: 700;
-        }
-    }
-
-    .div-btn-add-course {
-        width: 100%;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        margin-top: 0.25rem;
-
-        .add-btn {
-            display: inline-flex;
-            align-items: center;
-            gap: .6rem;
-            padding: .75rem 1.1rem;
-            border: none;
-            border-radius: 12px;
-            background: #e0e0e0;
-            box-shadow:
-                inset 5px 5px 10px #a6a6a6af,
-                inset -5px -5px 10px #ffffff,
-                0 6px 18px rgba(0, 0, 0, .06);
-            cursor: pointer;
-            transition: transform .25s ease, box-shadow .25s ease, background .25s ease;
-
-            .label {
-                color: #d4004b;
-                font-weight: 800;
-                white-space: nowrap;
-            }
-
-            .icon-wrap {
-                width: 22px;
-                height: 22px;
-                display: grid;
-                place-items: center;
-                border-radius: 50%;
-                transition: transform .3s ease, filter .3s ease;
-                filter: drop-shadow(0 0 4px rgba(255, 0, 120, .35));
-            }
-
-            .icon {
-                width: 100%;
-                height: 100%;
-                display: block;
-            }
-        }
-
-        .add-btn:hover {
-            transform: translateY(-2px);
-            background: linear-gradient(145deg, #f0f0f0, #cacaca);
-            box-shadow: 5px 5px 12px #b1b1b1, -5px -5px 12px #ffffff;
-
-            .icon-wrap {
-                transform: scale(1.12) rotate(6deg);
-            }
-        }
-
-        .add-btn .icon-wrap {
-            animation: pulse 1.8s ease-in-out infinite;
-        }
-
-        @keyframes pulse {
-
-            0%,
-            100% {
-                transform: scale(1);
-            }
-
-            50% {
-                transform: scale(1.06);
-            }
-        }
-
-        .paginator {
-            margin-top: 1.5rem;
-            border-radius: 10px;
-            background: linear-gradient(145deg, #f0f0f0, #cacaca);
-            box-shadow: 5px 5px 10px #b1b1b1, -5px -5px 10px #ffffff;
-        }
-
-    }
-
-    .paginator {
-        margin-top: 1.5rem;
-        border-radius: 10px;
-        background: linear-gradient(145deg, #f0f0f0, #cacaca);
-        box-shadow: 5px 5px 10px #b1b1b1,
-            -5px -5px 10px #ffffff;
-    }
-}
-
-@media (min-width: 768px) {
-    .courses {
-        flex-direction: row;
-    }
-}
-
-@media (min-width: 1100px) {
-    .courses-container {
-        .div-title-course {
-            margin-top: 8rem;
-        }
-    }
-}
